refactor(#163): rename testing subpaths to test-utils
