+++ projects/lib-contract-base/package.json
@@ -5,7 +5,7 @@
   "type": "module",
   "exports": {
     ".": "./src/index.ts",
-    "./testing": "./src/testing/index.ts"
+    "./test-utils": "./src/test-utils/index.ts"
   },
   "dependencies": {
     "@orpc/client": "1.14.6",
