+++ projects/lib-service-runtime/package.json
@@ -5,7 +5,7 @@
   "type": "module",
   "exports": {
     ".": "./src/index.ts",
-    "./testing": "./src/testing/index.ts"
+    "./test-utils": "./src/test-utils/index.ts"
   },
   "dependencies": {
     "@elysiajs/opentelemetry": "1.4.11",
